third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/cssom/cssstylesheet-insertrule-allowed.html

<!DOCTYPE html>
<html>
<head>
    <script src="/js-test-resources/js-test.js"></script>
    <meta http-equiv="Content-Security-Policy" content="style-src 'unsafe-inline' 'self' 'unsafe-eval'">
</head>
<body>
    <script>
        description("CSSStyleSheet::insertRule blocked if 'unsafe-eval' is not an allowed source of style.");
        shouldNotThrow("document.styleSheets[0].insertRule('x { display: none; }', 0);");
    </script>
</body>
</html>
